config: extract screen/IDS compilation to compiler_security_screen.go
Part B of #4121 (pure code motion). Moves the screen/IDS cluster
verbatim out of compiler_security.go: the screen threshold constants
and their commit-time advisory validators, plus compileScreen itself.

  - maxScanSweepThreshold, defaultSynFloodAttackThreshold, the Junos
    default flood/scan/sweep threshold block,
    synFloodSrcAttackRatioAdvisoryThreshold
  - validateScreenSynFloodSubThresholds (#3315 advisory band)
  - validateScreenScanSweepThresholds (#3230/#3317 clamp-warn)
  - compileScreen (the ids-option parser with the #3230 defaults,
    #3317 numeric gate, and #3318/#3332 unknown-leaf recording)

The constants and validators sit at the top of the original file while
compileScreen sits mid-file; their relative order is preserved in the
new module. Function bodies are byte-identical to origin/master. The
new file imports exactly fmt, math, sort, strconv (the union the moved
declarations use); math and sort drop out of the base file's import
block because only the screen cluster used them.

Validation: gofmt -l clean, go build ./pkg/config/ green.
